https://github.com/mrdoob/three.js/pull/14044#issuecomment-391117743

This PR is for improving `.updateMatrixWorld()` performance.

**The idea is**
- Introduce `.needsMatrixUpdate`.
- Set `.needsMatrixUpdate` `true` when `.position/quaternion/scale` is updated.
- Only when `.needsMatrixUpdate` is `true`, `.updateMatrixWorld()` calls `.updateMatrix()`.

In short, we don't call `.updateMatrix()` if we don't need to.

(I haven't tried lazy update yet)

**The changes are**
- Add `.onChange` to `Vector3`.
- Set `.needsMatrixUpdate` true in `onChange` callback of `Vector3/Quaternion/Euler`.
- Only when `.needsMatrixUpdate` is `true`, `.updateMatrixWorld()` calls `.updateMatrix()`.

The disadvantage of this idea is the code could be a bit complex. The purpose of this PR is to see how big/small changes are and how hard/easy to maintain, and plus, to see the benefit.

**UpdateMatrix benchmark Performance numbers. (`.updateMatrixWorld( true )` )**

On my Windows + Chrome

| graph | old [ops/sec] | new [ops/sec] | new / old | 
| ---- | ---- | ---- | ---- |
| depth=100, breadth=1 | 108746 | 136512 | 1.26 |
| depth=3, breadth=10 | 9784 | 11619 | 1.19 |
| depth=9, breadth=3 | 125 | 141 | 1.13 |

On my Windows + FireFox Nightly

| graph | old [ops/sec] | new [ops/sec] | new / old |
| ---- | ---- | ---- | ---- |
| depth=100, breadth=1 | 50817 | 68536 | 1.35 |
| depth=3, breadth=10 | 4633 | 7285 | 1.57 |
| depth=9, breadth=3 | 124 | 167 | 1.35 |

**UpdateMatrix benchmark Performance numbers. (`.updateMatrixWorld( undefined )` )**

Additional benefit is `.updataMatrixWorld( false or undefined )` can be much faster if local matrix and parent's world matrix aren't updated. Simply we can skip multiplication for world matrix calculation because both `.needsMatrixUpdate` and `.needsMatrixWorldUpdate` are `false`. 

On my Windows + Chrome

| graph | old [ops/sec] | new [ops/sec] | new / old | 
| ---- | ---- | ---- | ---- |
| depth=100, breadth=1 | 108746 | 542296 | 4.99 |
| depth=3, breadth=10 | 9784 | 53359 | 5.45 |
| depth=9, breadth=3 | 125 | 441 | 3.53 |

On my Windows + FireFox Nightly

| graph | old [ops/sec] | new [ops/sec] | new / old |
| ---- | ---- | ---- | ---- |
| depth=100, breadth=1 | 50817 | 345471 | 6.80 |
| depth=3, breadth=10 | 4633 | 48427 | 10.45 |
| depth=9, breadth=3 | 124 | 528 | 4.26 |