@takahirox 2018-05-24
More precisely, we can do this optimization in user code like

    // object creatation
    var object = new THREE.Object3D();
    object.matrixAutoUpdate = false;

    // animation
    object.position.x = newX;
    object.updateMatrix();

With this PR, this optimization will be automatically done.