@mquander 2018-05-25
I and my team are super grateful that you're working on this. Thanks.

This change makes me wonder what is the intended use case of the `force` parameter to `updateMatrixWorld`. There are a number of places in the three source which pass `true` where I would not have thought to pass it -- for example, in `Object3D.getWorldPosition` and `Object3D.getWorldQuaternion`. Changes like this would make it a bigger and bigger difference between the performance of passing `true` unnecessarily and not doing so. Is there any reason for callers to pass `true`, unless they have manually manipulated world matrices? If not, should places like this be changed?