@Mugen87 2018-05-25
Instead of introducing a new flag, I suggest a different approach similar to Babylon.js: Caching World Matrices.

https://books.google.de/books?id=6crECQAAQBAJ&pg=PA151&lpg=PA151&dq=babylons.js+cache+matrix&source=bl&ots=Jr6k7XRdMz&sig=8tOF53JDoT1OiK9BK9xOlrsaWJQ&hl=de&sa=X&ved=0ahUKEwilzayVvKDbAhVhEpoKHUnpCT0Q6AEITTAE#v=onepage&q=babylons.js%20cache%20matrix&f=false