@takahirox 2018-05-25
Would you explain a bit more?

I think introducing `.needsMatrixUpdate` is almost matrix/matrixWorld cache because they're calculated only when necessary.