@Mugen87 2018-05-25
I like the idea of the mentioned `isSynchronized()` method. So you keep information like `position`, `rotation` and `scale` in an internal cache and only return true if the cached data are equal with the real properties. In this case, you know that the cached matrix represents the actual transformation so you don't need a new calculation.