@takahirox 2018-05-25
Like this?

    Object.prototype.isSynchronized = function () {
        return (
            this.position.x === this.cachedPosition.x &&
            this.position.y === this.cachedPosition.y &&
            this.position.z === this.cachedPosition.z &&
            this.quaternion.x === this.cachedQuaternion.x &&
            this.quaternion.y === this.cachedQuaternion.y &&
            this.quaternion.z === this.cachedQuaternion.z &&
            this.quaternion.w === this.cachedQuaternion.w &&
            this.scale.x === this.cachedScale.x &&
            this.scale.y === this.cachedScale.y &&
            this.scale.z === this.cachedScale.z
        )
    }