@WestLangley 2018-05-25
@takahirox 

It is great you are interested in this issue.

`updateMatrixWorld()` was written to be applied to a _scene_, and traverses the scene graph in an efficient manner. It is not the correct method to be used in the case of scene _node_.

For example,
```javascript
object.parent.rotation.x += 1;
object.updateMatrixWorld();
```
This is a long-standing issue. See https://github.com/mrdoob/three.js/pull/6448#issuecomment-95624064.

It may be beneficial to have a look at the history of this topic.