@gkjohnson 2018-05-26
Hey @talahirox!

It's great to see that someone is working on this. I was thinking about looking into matrix update optimizations, as well! I'd be happy to help if you're looking for someone to contribute.

I'm sure you already have some of this in mind (and have a good portion of it in, already), but here's how I was thinking about the optimization:

#### Assumptions
- If an objects localMatrix changes, then the worldMatrix needs updating.
- If an objects worldMatrix changes, then all of the childrens (and childrens childrens) world matrices needs updating.
- If an objects worldMatrix is being updated, then its parents worldMatrix needs to be made up to date.

#### Implementation
- If an objects localMatrix needs updating, then the worldMatrix should be marked as needing an update.
- If the worldMatrix is marked as needing an update, then mark all children as needing their world matrices updated.
- If an objects worldMatrix is already marked as needing an update, then all of its children should already be marked.
- If a call to "updateWorldMatrix" is made, then it should call its parent's "updateWorldMatrix" function to.
- The "updateWorldMatrix" function doesn't need to run if the world matrix does not need an update.

In my opinion, this should simplify things a lot! It means that you should only have to call "updateWorldMatrix" on a child and know immediately whether or not it's up to date. This also means that the renderer will only have to call "updateWorldMatrix" on the meshes it wants to update rather than updating the full tree (or nodes that don't even need to be updated). 

@Mugen87 I didn't know Babylon did that. Sounds like it would optimize for the case where you move something and then it moves back, but wouldn't you still have to check everything to see if it has needs to be updated? This approach would let you queue up only the things that need to be changed (rather than iterating over everything). I be misunderstanding, though.

A couple other related thoughts:
- Adding the ability to register for when something has had its transform changed or marked as changed would be nice (and would enable some nice new extensions to THREE)
- A similar approach could be used to more optimally create and update bounding boxes on internal tree nodes, which would be nice to have, as well.