@pailhead 2018-05-29
Could we maybe consider some kind of an API where we would have `onBeforeUpdate` and `onAfterUpdate`. I suggested this back when `onBeforeRender` got accepted. 

I did not dive too deep into this, but it seems like it would make sense. Before would be a time where one sets `myMesh.position`, after would be where you have the matrices computed. Is there a pattern here that could be established?