@takahirox 2018-06-07
I tried a cache approach. Much less changes but less performance gain. Roughly the performance gain is half or more less from this PR. I think 10 times comparing `positioin.xyz, quaternion.xyzw, scale.xyz` isn't so fast.

    updateMatrix: function () {

        if ( ! this.position.equals( this.positionCache ) ||
            ! this.quaternion.equals( this.quaternionCache ) ||
            ! this.scale.equals( this.scaleCache ) ) {

            this.matrix.compose( this.position, this.quaternion, this.scale );

            this.positionCache.copy( this.position );
            this.quaternionCache.copy( this.quaternion );
            this.scaleCache.copy( this.scale );

            this.matrixWorldNeedsUpdate = true;

        }

    },

Which one do you prefer? I have two options in my mind. First is this PR because I wanna improve the performance of the hot-spot function as much as possible. Second one is the cache approach. If user want further optimization user can set `object.matrixAutoUpdate = false;` and call `object.updateMatrix();` when necessary by themselves.