@Mugen87 2018-06-07
I vote for the cache approach.