@takahirox 2018-06-07
@gkjohnson 

Thanks for sharing your idea. I wanna focus on the current behavior of `.updateMatrixWorls()` so far which is 

> to be applied to a scene, and traverses the scene graph in an efficient manner.

as @WestLangley explained.

I'd like to discuss in another thread for calculating world matrix of a node.