@takahirox 2018-06-07
> I vote for the cache approach.

How about adopting the caching approach now because the change is very small, and we'll consider further optimization like this PR idea when users request further improvement?

Any thoughts? Or any better solutions?

/cc @mrdoob